Renames donations folder as donorschoose, DonationController as BotController
