--- app/lib/donations/package.json
@@ -1,4 +1,0 @@
-{
-  "name" : "ds-donations",
-  "main" : "./donations.js"
-}
